Move helpers into __main__.
