+++ ptvsd/debugger.py
@@ -2,56 +2,20 @@
 # Licensed under the MIT License. See LICENSE in the project root
 # for license information.
 
-import sys
-
-# import the wrapper first, so that it gets a chance
-# to detour pydevd socket functionality.
-import ptvsd.wrapper
+from ptvsd.__main__ import run_module, run_file
 
 
 __author__ = "Microsoft Corporation <ptvshelp@microsoft.com>"
 __version__ = "4.0.0a5"
 
+# TODO: not needed?
 DONT_DEBUG = []
 
 
-def debug(filename, port_num, debug_id, debug_options, run_as):
+def debug(filename, port_num, debug_id, debug_options, run_as, **kwargs):
     # TODO: docstring
     address = (None, port_num)
     if run_as == 'module':
-        _run_module(address, filename)
+        run_module(address, filename, **kwargs)
     else:
-        _run_file(address, filename)
-
-
-def _run_module(address, modname):
-    filename = modname + ':'
-    argv = _run_argv(address, filename)
-    argv.insert(argv.index('--file'), '--module')
-    _run(argv)
-
-
-def _run_file(address, filename):
-    argv = _run_argv(address, filename)
-    _run(argv)
-
-
-def _run_argv(address, filename):
-    host, port = address
-    if host is None:
-        host = '127.0.0.1'
-    return [
-        '--port', str(port),
-        '--client', host,
-        '--file', filename,
-    ]
-
-
-def _run(argv):
-    pydevd = ptvsd.wrapper.install(argv)
-    sys.argv[1:0] = argv
-    try:
-        pydevd.main()
-    except SystemExit as ex:
-        ptvsd.wrapper.ptvsd_sys_exit_code = int(ex.code)
-        raise
+        run_file(address, filename, **kwargs)
